Refactor Remote DTOs

+++ remote/src/main/kotlin/com/rainbow/remote/dto/RemoteComment.kt
@@ -4,6 +4,6 @@ data class RemoteComment(
     val id: String?,
     val author: String?,
     val edited: String?,
-    val ups: Int?,
-    val downs: Int?,
+    val ups: Long?,
+    val downs: Long?,
 )
